requirements.txt
uvicorn==0.17.6
fastap==0.75.2
pandas==1.4.2
SQLAlchemy==1.4.36